Clean code to remove compilation warnings

+++ src/init.c
@@ -1,4 +1,5 @@
 #include <drivers/dma.h>
+#include <stdio.h>
 
 void cam_dma_user_callback(const struct device *dma_dev, void *arg,
 			      uint32_t id, int error_code)
@@ -37,9 +38,9 @@ void dma_init_cams() {
 	/*set block size, driver will get image width from that */
 	dma_block_cfg_cam.block_size = 800 * 600;
 
-	dma_block_cfg_cam.dest_address = &img_buff_1;
+	dma_block_cfg_cam.dest_address = (uint32_t)&img_buff_1;
 	dma_config(fastvdma_dev_cam_1, 0, &dma_cfg_cam);
-	dma_block_cfg_cam.dest_address = &img_buff_2;
+	dma_block_cfg_cam.dest_address = (uint32_t)&img_buff_2;
 	dma_config(fastvdma_dev_cam_2, 0, &dma_cfg_cam);
 }
